Move mapper classes in Mapper namespace
